Adding refactored code

+++ cnnhw.py
@@ -148,8 +148,36 @@ def get_dataset():
 
   return X_train, y_train, X_test, y_test
 
+def make_or_restore_model(cdir, checkpoints_exist=False):
+  # Either restore the latest model, or create a fresh one
+  # if there is no checkpoint available.
+  if checkpoints_exist:
+    checkpoints = [cdir + "/" + name for name in os.listdir(cdir)]
+    if checkpoints:
+        latest_checkpoint = max(checkpoints, key=os.path.getctime)
+        print("Restoring from", latest_checkpoint)
+        return keras.models.load_model(latest_checkpoint)
+    print("Creating a new model")
+  return get_compiled_model()
 
+def train_model(cdir, model, X_train, y_train, epochs):
+  # Directory for TensorBoard logs
+  log_dir = tdir + "/logs/fit/" + datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
 
+  # Create a callback for model checkpoints
+  checkpoint_cb = ModelCheckpoint(filepath=cdir + "/ckpt-{epoch}", save_freq="epoch")
+
+  # Create a TensorBoard callback
+  tensorboard_callback = TensorBoard(log_dir=log_dir, histogram_freq=1)
+
+  # Train the model
+  model.fit(X_train, y_train, epochs=epochs, callbacks=[checkpoint_cb, tensorboard_callback])
+
+def get_data_and_model(cdir, checkpoint_exists):
+  X_train, y_train, X_test, y_test = get_dataset()
+
+  model = make_or_restore_model(cdir, checkpoint_exists)
+  return X_train, y_train, X_test, y_test, model
 
 if __name__ == "__main__":
   argv = sys.argv
@@ -204,57 +232,30 @@ if __name__ == "__main__":
   else:
     cdir = "/tmp/" + user + "/ckpt"
     tdir = "/tmp/" + user +"/tb"
-  
-  # Remove the directories and any files within and Create the directories if all checkpoints are present
-  
+
+  checkpoint_exists = True
+
+  epochs = 15
+
   if (os.path.exists(cdir)):
-    os.system(f"rm -rf {cdir}")
-  if (os.path.exists(tdir)): 
-    os.system(f"rm -rf {tdir}")
-  
+    files = files = os.listdir(cdir)
+    # if there is a file for every epoch of training this means training is done
+    # and we need to remove them all
+    if all(f'ckpt-{i}' in files for i in range(1, epochs + 1)):
+      os.system(f"rm -rf {cdir}")
+      os.system(f"rm -rf {tdir}")
+      checkpoint_exists = False
+  else:
+    checkpoint_exists = False
+  # we'll make sure these directories exist
   os.makedirs(cdir, exist_ok=True)
   os.makedirs(tdir, exist_ok=True)
-
-  def make_or_restore_model(checkpoints_exist=False):
-    # Either restore the latest model, or create a fresh one
-    # if there is no checkpoint available.
-    if checkpoints_exist:
-      checkpoints = [cdir + "/" + name for name in os.listdir(cdir)]
-      if checkpoints:
-          latest_checkpoint = max(checkpoints, key=os.path.getctime)
-          print("Restoring from", latest_checkpoint)
-          return keras.models.load_model(latest_checkpoint)
-      print("Creating a new model")
-    return get_compiled_model()
-  
-  def get_data_and_model():
-    X_train, y_train, X_test, y_test = get_dataset()
-    
-    model = make_or_restore_model(checkpoint_exists)
-    return X_train, y_train, X_test, y_test, model
-
-
-
-
-
-  def train_model(model, X_train, y_train):
-    # Directory for TensorBoard logs
-    log_dir = tdir + "/logs/fit/" + datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
-
-    # Create a callback for model checkpoints
-    checkpoint_cb = ModelCheckpoint(filepath=cdir + "/ckpt-{epoch}", save_freq="epoch")
-
-    # Create a TensorBoard callback
-    tensorboard_callback = TensorBoard(log_dir=log_dir, histogram_freq=1)
-
-    # Train the model
-    model.fit(X_train, y_train, epochs=15, callbacks=[checkpoint_cb, tensorboard_callback])
 
   strategy = tf.distribute.experimental.MultiWorkerMirroredStrategy()
 
   with strategy.scope():
-    X_train, y_train, X_test, y_test, model = get_data_and_model()
-    train_model(model, X_train, y_train)
+    X_train, y_train, X_test, y_test, model = get_data_and_model(checkpoint_exists)
+    train_model(cdir, model, X_train, y_train, epochs)
 
     # Model evaluation and prediction
     test_loss, test_accuracy = model.evaluate(X_test, y_test)
